# Fetch table metadata in parallel and asynchronously and add them to a queue for faster processing